keep header and sidepanel during navigation
