+++ apps/statgpt-admin-frontend/src/app/(main)/audit-logs/page.tsx
@@ -2,12 +2,11 @@ import { cookies, headers } from 'next/headers';
 import { redirect } from 'next/navigation';
 
 import { AuditLogsListView } from '@/src/components/AuditLogs/AuditLogsListView';
-import { MainShell } from '@/src/components/MainShell/MainShell';
-import { NoAccessView } from '@/src/components/NoAccess/NoAccessView';
+import { ForbiddenTrigger } from '@/src/components/NoAccess/ForbiddenTrigger';
 import { SIGN_IN_LINK } from '@/src/constants/auth';
 import { getIsInvalidSession, getUserToken } from '@/src/utils/auth/get-token';
 import { getIsEnableAuthToggle } from '@/src/utils/get-auth-toggle';
-import { auditLogsApi } from '../api/api';
+import { auditLogsApi } from '../../api/api';
 import { logger } from '@/src/server/logger';
 
 export const dynamic = 'force-dynamic';
@@ -26,7 +25,7 @@ export default async function Page() {
   if (enumsResult.ok) {
     enums = enumsResult.data;
   } else if (enumsResult.error.status === 403) {
-    return <NoAccessView />;
+    return <ForbiddenTrigger />;
   } else {
     logger.error(
       `Getting audit log enum values error ${enumsResult.error.message}`,
@@ -34,11 +33,9 @@ export default async function Page() {
   }
 
   return (
-    <MainShell>
-      <AuditLogsListView
-        enums={enums}
-        initialError={enumsResult.ok ? null : enumsResult.error.message}
-      />
-    </MainShell>
+    <AuditLogsListView
+      enums={enums}
+      initialError={enumsResult.ok ? null : enumsResult.error.message}
+    />
   );
 }
