+++ apps/statgpt-admin-frontend/src/app/(main)/channels/page.tsx
@@ -1,14 +1,13 @@
 import { cookies, headers } from 'next/headers';
 import { redirect } from 'next/navigation';
 
-import { dataSourcesApi } from '@/src/app/api/api';
+import { channelsApi } from '@/src/app/api/api';
 import { ListView } from '@/src/components/ListView/ListView';
-import { MainShell } from '@/src/components/MainShell/MainShell';
-import { NoAccessView } from '@/src/components/NoAccess/NoAccessView';
+import { ForbiddenTrigger } from '@/src/components/NoAccess/ForbiddenTrigger';
 import { SIGN_IN_LINK } from '@/src/constants/auth';
-import { DATA_SOURCE_COLUMNS_WITH_ACTIONS } from '@/src/constants/columns/grid-columns';
+import { CHANNELS_COLUMNS } from '@/src/constants/columns/grid-columns';
 import { Menu } from '@/src/constants/menu';
-import { DataSource } from '@/src/models/data-source';
+import { Channel } from '@/src/models/channel';
 import { RequestData } from '@/src/models/request-data';
 import { logger } from '@/src/server/logger';
 import { getIsInvalidSession, getUserToken } from '@/src/utils/auth/get-token';
@@ -25,26 +24,24 @@ export default async function Page() {
     return redirect(SIGN_IN_LINK);
   }
 
-  let data = { data: [] as DataSource[] } as RequestData<DataSource> | null;
+  let data = { data: [] as Channel[] } as RequestData<Channel> | null;
 
-  const result = await dataSourcesApi.getDataSources(token);
+  const result = await channelsApi.getChannels(token);
   if (result.ok) {
     data = result.data;
   } else if (result.error.status === 403) {
-    return <NoAccessView />;
+    return <ForbiddenTrigger />;
   } else {
-    logger.error(`Getting data sources error ${result.error.message}`);
+    logger.error(`Getting channels error ${result.error.message}`);
   }
 
   return (
-    <MainShell>
-      <ListView
-        menuItem={Menu.DATA_SOURCES}
-        colDefs={DATA_SOURCE_COLUMNS_WITH_ACTIONS}
-        data={data?.data || []}
-        emptyDataTitle="No Data Sources"
-        initialError={result.ok ? null : result.error.message}
-      />
-    </MainShell>
+    <ListView
+      menuItem={Menu.CHANNELS}
+      colDefs={CHANNELS_COLUMNS}
+      data={data?.data || []}
+      emptyDataTitle="No Channels"
+      initialError={result.ok ? null : result.error.message}
+    />
   );
 }
