+++ apps/statgpt-admin-frontend/src/app/(main)/data-sets/page.tsx
@@ -1,20 +1,19 @@
-'use server';
-
 import { cookies, headers } from 'next/headers';
 import { redirect } from 'next/navigation';
 
-import { documentsApi } from '@/src/app/api/api';
+import { dataSetsApi } from '@/src/app/api/api';
 import { ListView } from '@/src/components/ListView/ListView';
-import { MainShell } from '@/src/components/MainShell/MainShell';
-import { NoAccessView } from '@/src/components/NoAccess/NoAccessView';
+import { ForbiddenTrigger } from '@/src/components/NoAccess/ForbiddenTrigger';
 import { SIGN_IN_LINK } from '@/src/constants/auth';
-import { DOCUMENTS_COLUMNS_WITH_ACTIONS } from '@/src/constants/columns/grid-columns';
+import { DATA_SETS_COLUMNS_WITH_ACTIONS } from '@/src/constants/columns/grid-columns';
 import { Menu } from '@/src/constants/menu';
-import { Document } from '@/src/models/document';
+import { DataSet } from '@/src/models/data-sets';
 import { RequestData } from '@/src/models/request-data';
 import { logger } from '@/src/server/logger';
 import { getIsInvalidSession, getUserToken } from '@/src/utils/auth/get-token';
 import { getIsEnableAuthToggle } from '@/src/utils/get-auth-toggle';
+
+export const dynamic = 'force-dynamic';
 
 export default async function Page() {
   const isEnableAuth = getIsEnableAuthToggle();
@@ -25,26 +24,24 @@ export default async function Page() {
     return redirect(SIGN_IN_LINK);
   }
 
-  let data = { data: [] as Document[] } as RequestData<Document> | null;
+  let data = { data: [] as DataSet[] } as RequestData<DataSet> | null;
 
-  const result = await documentsApi.getList(null);
+  const result = await dataSetsApi.getDataSets(token);
   if (result.ok) {
     data = result.data;
   } else if (result.error.status === 403) {
-    return <NoAccessView />;
+    return <ForbiddenTrigger />;
   } else {
-    logger.error(`Getting documents error ${result.error.message}`);
+    logger.error(`Getting data sets error ${result.error.message}`);
   }
 
   return (
-    <MainShell>
-      <ListView
-        menuItem={Menu.DOCUMENTS}
-        colDefs={DOCUMENTS_COLUMNS_WITH_ACTIONS}
-        data={(data?.results as any[]) || []}
-        emptyDataTitle="No Documents"
-        initialError={result.ok ? null : result.error.message}
-      />
-    </MainShell>
+    <ListView
+      menuItem={Menu.DATA_SETS}
+      colDefs={DATA_SETS_COLUMNS_WITH_ACTIONS}
+      data={data?.data || []}
+      emptyDataTitle="No Datasets"
+      initialError={result.ok ? null : result.error.message}
+    />
   );
 }
