+++ apps/statgpt-admin-frontend/src/app/(main)/layout.tsx
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+
+import { MainShell } from '@/src/components/MainShell/MainShell';
+
+export default function MainLayout({ children }: { children: ReactNode }) {
+  return <MainShell>{children}</MainShell>;
+}
